+++ src/server/http/responseDecorators/defaultDecorators.js
@@ -1,0 +1,7 @@
+import noUndefinedDecorator from './noUndefinedDecorator'
+import toStringDecorator from './toStringDecorator'
+
+export default [
+  toStringDecorator,
+  noUndefinedDecorator
+]