+++ src/server/http/responseDecorators/noUndefinedDecorator.js
@@ -1,0 +1,6 @@
+export default function noUndefinedDecorator (response, name) {
+  if (response === void 0) {
+    throw new TypeError(`${name} must return a response or throw an error`)
+  }
+  return response
+}